Remove unecesarry documentation.
