+++ config/java.d/jmx.conf
@@ -15,6 +15,7 @@
     "jmxServers": [
         // Example configuration for a java program started on the same host with these flags:
         // -Dcom.sun.management.jmxremote.port=9999 -Dcom.sun.management.jmxremote.authenticate=false -Dcom.sun.management.jmxremote.ssl=false
+        /*
         {
             // Name displayed at the dashboard menu.
             "name": "ExampleJMX",
@@ -23,6 +24,7 @@
             // @see: https://docs.oracle.com/cd/E19159-01/819-7758/gcnqf/index.html
             "serviceUrl": "service:jmx:rmi:///jndi/rmi://:9999/jmxrmi"
         }
+        */
     ],
     // Global chart configurations.
     // Every monitored JMX Servers tries to monitor each chart configuration.
@@ -60,6 +62,7 @@
                     // M(X)Bean attribute to collect.
                     // The getter of the attribute must return one of these types.
                     // - Integer (int)
+                    // - Double (double)
                     // - Long (long)
                     "value": "ProcessCpuTime",
                     // Name at the dashboard
@@ -248,4 +251,4 @@
             ]
         }
     ]
-} // 
+}
