+++ src/main/java/org/firehol/netdata/entity/ChartType.java
@@ -20,7 +20,7 @@ package org.firehol.netdata.entity;
 
 public enum ChartType {
 	LINE, AREA, STACKED;
-	
+
 	@Override
 	public String toString() {
 		return name().toLowerCase();
