+++ src/main/java/org/firehol/netdata/exception/InitializationException.java
@@ -20,8 +20,6 @@ package org.firehol.netdata.exception;
 
 /**
  * Wraps an Exception during initialization. Marks the object as invalid.
- * 
- * @author Simon Nagl
  */
 public class InitializationException extends Exception {
 	private static final long serialVersionUID = 6446433868518776741L;
