+++ src/main/java/org/firehol/netdata/exception/UnreachableCodeException.java
@@ -18,13 +18,6 @@
 
 package org.firehol.netdata.exception;
 
-/**
- * Thrown to indicate that code has been reached which the author did not think
- * it could be reached.
- * 
- * @author Simon Nagl
- * @since 1.0
- */
 public class UnreachableCodeException extends RuntimeException {
 	private static final long serialVersionUID = 9126523210158499016L;
