+++ src/main/java/org/firehol/netdata/plugin/jmx/configuration/JmxDimensionConfiguration.java
@@ -24,9 +24,6 @@ import lombok.Setter;
 /**
  * Configuration scheme of a dimension of a chart created by the
  * {@link org.firehol.netdata.plugin.jmx.JmxPlugin}.
- * 
- * @since 1.0.0
- * @author Simon Nagl
  */
 @Getter
 @Setter
