+++ src/main/java/org/firehol/netdata/plugin/jmx/configuration/JmxPluginConfiguration.java
@@ -28,6 +28,7 @@ import lombok.Setter;
  * Configuration scheme to configure
  * {@link org.firehol.netdata.plugin.jmx.JmxPlugin}
  * 
+ * @since 1.0.0
  * @author Simon Nagl
  */
 @Getter
@@ -45,5 +46,14 @@ public class JmxPluginConfiguration {
 	 */
 	private List<JmxServerConfiguration> jmxServers = new ArrayList<>();
 
+	/**
+	 * A list of chart configurations.
+	 * 
+	 * <p>
+	 * Every monitored JMX Servers tries to monitor each chart in this list. If a
+	 * JMX Server does not have the required M(X)Beans we won't try adding it over
+	 * and over again.
+	 * </p>
+	 */
 	private List<JmxChartConfiguration> commonCharts = new ArrayList<>();
 }
