+++ src/main/java/org/firehol/netdata/plugin/jmx/exception/JmxMBeanServerConnectionException.java
@@ -18,12 +18,6 @@
 
 package org.firehol.netdata.plugin.jmx.exception;
 
-/**
- * Represents exceptions while initializing a connection to a JxmMBean Server.
- * 
- * @since 1.0.0
- * @author Simon Nagl
- */
 public class JmxMBeanServerConnectionException extends JmxPluginException {
 	private static final long serialVersionUID = -6153969842214336278L;
