--- src/main/java/org/firehol/netdata/utils/Converter.java
@@ -1,33 +1,0 @@
-/*
- * Copyright (C) 2017 Simon Nagl
- *
- * netadata-plugin-java-daemon is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- *
- */
-
-package org.firehol.netdata.utils;
-
-/**
- * A converter converts an object of class S to an object of class D.
- * 
- * @author Simon Nagl
- *
- * @param <S>
- *            Source
- * @param <D>
- *            Destination
- */
-public interface Converter<S, D> {
-	D convert(S source);
-}
